+++ subsys/bluetooth/controller/ll_sw/lll_df.h
@@ -6,3 +6,6 @@
 
 int lll_df_init(void);
 int lll_df_reset(void);
+
+/* Provides number of available antennae for Direction Finding */
+uint8_t lll_df_ant_num_get(void);
