+++ subsys/bluetooth/controller/ll_sw/openisa.cmake
@@ -30,9 +30,6 @@ if(CONFIG_BT_LL_SW_SPLIT)
       )
     endif()
   endif()
-    zephyr_library_include_directories(
-      ll_sw/openisa/lll
-    )
   zephyr_library_sources_ifdef(
     CONFIG_BT_CTLR_DTM
     ll_sw/openisa/lll/lll_test.c
