+++ subsys/bluetooth/controller/ll_sw/ull_conn_iso.c
@@ -11,6 +11,7 @@
 
 #include "lll.h"
 #include "lll_conn_iso.h"
+
 #include "ull_conn_iso_types.h"
 
 #define BT_DBG_ENABLED IS_ENABLED(CONFIG_BT_DEBUG_HCI_DRIVER)
