+++ subsys/bluetooth/controller/ll_sw/ull_sched.c
@@ -20,7 +20,7 @@
 #include "pdu.h"
 
 #include "lll.h"
-#include "lll_vendor.h"
+#include "lll/lll_vendor.h"
 #include "lll_scan.h"
 #include "lll_conn.h"
