+++ subsys/bluetooth/controller/ll_sw/ull_slave.c
@@ -21,16 +21,16 @@
 #include "ticker/ticker.h"
 
 #include "pdu.h"
-#include "ll.h"
 
 #include "lll.h"
-#include "lll_vendor.h"
 #include "lll_clock.h"
+#include "lll/lll_vendor.h"
+#include "lll/lll_adv_types.h"
 #include "lll_adv.h"
+#include "lll/lll_adv_pdu.h"
 #include "lll_conn.h"
 #include "lll_slave.h"
 #include "lll_filter.h"
-#include "lll_tim_internal.h"
 
 #include "ull_adv_types.h"
 #include "ull_conn_types.h"
@@ -40,6 +40,8 @@
 #include "ull_adv_internal.h"
 #include "ull_conn_internal.h"
 #include "ull_slave_internal.h"
+
+#include "ll.h"
 
 #define BT_DBG_ENABLED IS_ENABLED(CONFIG_BT_DEBUG_HCI_DRIVER)
 #define LOG_MODULE_NAME bt_ctlr_ull_slave
